+++ components/pedal-card/pedal-card-styled
@@ -1,0 +1,1 @@
+import styled
