+++ components/pedal-gallery/index.js
@@ -1,0 +1,16 @@
+import React from "react";
+import PedalCardMain from "../pedal-card/index";
+import { Pedals } from "../../lib/pedalfx-data";
+import { ShowCase } from "./";
+
+const PedalGallery = () => {
+  return (
+    <ShowCase>
+      {Pedals.map((pedalData, id) => (
+        <PedalCardMain key={id} pedalData={pedalData} />
+      ))}
+    </ShowCase>
+  );
+};
+
+export default PedalGallery;
