--- pages/components/pedalfx-data.js
@@ -1,188 +1,0 @@
-const Pedals = () => {
-  const pedals = [
-    {
-      id: "1",
-      NAME: "Spacetime V2",
-      MANUFACTURER: "Vahlbruch",
-      MADE_IN: "Germany",
-      CATEGORY: "Delay",
-      MONO_STEREO: "mono",
-      BYPASS: "true OR buffered",
-      VOLTAGE: "9V",
-      POLARITY: "center negative",
-      INPUT_IMPEDANCE: "1.0 M Ohm",
-      OUTPUT_IMPEDANCE: "1.0 k Ohm",
-      BYPASS_CURRENT_DRAW__manufacturer: " mA",
-      ENGAGED_CURRENT_DRAW_manufacturer: "60 mA",
-      BYPASS_CURRENT_DRAW_own_measurement: " mA",
-      ENGAGED_CURRENT_DRAW_own_measurement: " mA",
-      WIDTH_mm: "70",
-      DEPTH_mm: "115",
-      HEIGHT_mm: "48",
-      PICTURE: " ./img/Spacetime.png",
-      FOOTSWITCH_1: "on/ off",
-      FOOTSWITCH_2: "",
-      FOOTSWITCH_3: "",
-      FOOTSWITCH_4: "",
-      KNOB_1: "MIX",
-      KNOB_2: "TIME",
-      KNOB_3: "REPEATS",
-      KNOB_4: "",
-      KNOB_5: "",
-      KNOB_6: "",
-      KNOB_7: "",
-      KNOB_8: "",
-      SWITCH_1: "bright - mellow - dark",
-      SWITCH_2: "",
-      SWITCH_3: "",
-      SWITCH_4: "",
-      SWITCH_5: "",
-      SWITCH_6: "",
-      SWITCH_7: "",
-      SWITCH_8: "",
-      INTERNAL_DIPSWITCH_1: "TBP/ BBP",
-      INTERNAL_DIPSWITCH_2: "kill dry",
-      POWER_JACK: ["2.1 mm", " TOP"],
-      IN: "right",
-      OUT: "left",
-      EXP_In: "",
-      SOUND_DESCRIPTION:
-        "I like the warm repeats they are just lying behind the signal and give a beatiful ambiendce.",
-      MY_COMMENT: "AWWWWWSOME DELAY AY AY AY",
-      NICE_TO_KNOW:
-        "You can choose if the pedal powers up in on or bypass state. To change the state: Step 1: Unplug the power supply cable. Step 2: Hold down on/off MagTraB button. Step 3: Plug in again the power supply cable. Step 4: Release on/off MagTraB button. DONE! The next time your pedal powers up it will automatically start in the other state.",
-      TAGS: ["#warm", " #slapback", " #alwayson", " #timebased"],
-    },
-    {
-      id: "2",
-      NAME: "DELTA II",
-      MANUFACTURER: "Spaceman",
-      CATEGORY: ["Tremolo", " Harmonic Trem"],
-      MADE_IN: "USA",
-      MONO_STEREO: "mono",
-      BYPASS: "true",
-      VOLTAGE: "9 - 18 V",
-      POLARITY: "center negative",
-      INPUT_IMPEDANCE: "1.0 M Ohm",
-      OUTPUT_IMPEDANCE: "",
-      BYPASS_CURRENT_DRAW__manufacturer: "19.5 mA",
-      ENGAGED_CURRENT_DRAW_manufacturer: "45 mA",
-      BYPASS_CURRENT_DRAW_own_measurement: " mA",
-      ENGAGED_CURRENT_DRAW_own_measurement: " mA",
-      WIDTH_mm: "66",
-      DEPTH_mm: "120",
-      HEIGHT_mm: "55",
-      PICTURE: " ./img/Delta2.png",
-      FOOTSWITCH_1: "on/ off",
-      FOOTSWITCH_2: "",
-      FOOTSWITCH_3: "",
-      FOOTSWITCH_4: "",
-      KNOB_1: "DEPTH",
-      KNOB_2: "FREQ",
-      KNOB_3: "RATE",
-      KNOB_4: "LAG",
-      KNOB_5: "RES",
-      KNOB_6: "",
-      KNOB_7: "",
-      KNOB_8: "",
-      SWITCH_1: "VOICE",
-      SWITCH_2: "",
-      SWITCH_3: "",
-      SWITCH_4: "",
-      SWITCH_5: "",
-      SWITCH_6: "",
-      SWITCH_7: "",
-      SWITCH_8: "",
-      INTERNAL_DIPSWITCH_1: "EXP_parameter",
-      INTERNAL_DIPSWITCH_2: "",
-      POWER_JACK: ["2.1 mm", " left"],
-      IN: "right",
-      OUT: "left",
-      EXP_In: "right",
-      SOUND_DESCRIPTION:
-        "It is hard to tame the Rate that gets fast quite quickly, but once you tame it, this is classic to fun and back again.",
-      MY_COMMENT: "Fun of a Trem",
-      NICE_TO_KNOW:
-        "ALTERNATE MODES - Apply power with the footswitch depressed. The LAG control transforms into a 4 position mode selector with additional LFO presets. Mode 1 - LFO output 1 runs at normal speed, and LFO output 2 at half speed. Mode 2 - Same as mode 1, but offbeat from each other. Mode 3 - LFO output 2 runs at normal speed, and LFO output 1 at half speed. Mode 4 - Same as mode 3, but offbeat from each other.",
-      TAGS: ["#harmonic", "# versatile", "# expressionn", "# modulation"],
-    },
-    {
-      id: "3",
-      NAME: "MUDHONEY",
-      MANUFACTURER: "T-Rex",
-      CATEGORY: ["Overdrive", "Distortion"],
-      MADE_IN: "Denmark",
-      MONO_STEREO: "mono",
-      BYPASS: "true",
-      VOLTAGE: "8.5 - 12.5 V",
-      POLARITY: "center negative",
-      INPUT_IMPEDANCE: "> 1 M Ohm",
-      OUTPUT_IMPEDANCE: "< 1 k Ohm",
-      BYPASS_CURRENT_DRAW__manufacturer: "",
-      ENGAGED_CURRENT_DRAW_manufacturer: "6 mA",
-      BYPASS_CURRENT_DRAW_own_measurement: " mA",
-      ENGAGED_CURRENT_DRAW_own_measurement: " mA",
-      WIDTH_mm: "100",
-      DEPTH_mm: "120",
-      HEIGHT_mm: "55",
-      PICTURE: " ./img/Mudhoney.png",
-      FOOTSWITCH_1: "on/ off",
-      FOOTSWITCH_2: "",
-      FOOTSWITCH_3: "",
-      FOOTSWITCH_4: "",
-      KNOB_1: "GAIN",
-      KNOB_2: "LEVEL",
-      KNOB_3: "TONE",
-      KNOB_4: "",
-      KNOB_5: "",
-      KNOB_6: "",
-      KNOB_7: "",
-      KNOB_8: "",
-      SWITCH_1: "BOOST",
-      SWITCH_2: "",
-      SWITCH_3: "",
-      SWITCH_4: "",
-      SWITCH_5: "",
-      SWITCH_6: "",
-      SWITCH_7: "",
-      SWITCH_8: "",
-      INTERNAL_DIPSWITCH_1: "",
-      INTERNAL_DIPSWITCH_2: "",
-      POWER_JACK: ["2.1 mm", " top"],
-      IN: "top",
-      OUT: "top",
-      EXP_In: "",
-      SOUND_DESCRIPTION:
-        "You can cover a lot of tones with this one. Crunch to Fuzz, it is at it is.",
-      MY_COMMENT:
-        "Find a good compromise setting and change drastically by using the boost switch",
-      NICE_TO_KNOW: "",
-      TAGS: ["#crunch", " #fuzzy", " #distortion"],
-    },
-  ];
-
-  return (
-    <div>
-      {pedals.map((pedal) => (
-        <div key={pedal.id}>
-          <h3>{pedal.NAME}</h3>
-          <ul>
-            {Object.entries(pedal).map(([key, value]) => {
-              if (value) {
-                return (
-                  <li key={key}>
-                    {key}: {value}
-                  </li>
-                );
-              }
-              return null;
-            })}
-          </ul>
-        </div>
-      ))}
-    </div>
-  );
-};
-export default Pedals;
-
-console.log(Pedals);
