+++ pages/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PedalGallery from "./components/pedal-gallery";
+import PedalGallery from "../components/pedal-gallery/index";
 
 export default function HomePage() {
   return (
